Main: 修改Java/main的createtable指令其中notify_center追加通知時間欄位、通知類別改為TinyInt、讀取狀態預設值為0 修改Java/main的假資料指令其中notify_center追加通知時間資料並刪除讀取狀態 修改NotifyCenter類別的的通知類別型別為Byte 新增NotifyCenter Service、HibernateDAO、Controller、讀取狀態列舉、通知類型列舉 實作後台系統通知功能：列表(查全部)、詳情(查單一)
minorfix:
修改部分類別名稱及simpledatetable js檔案字串中文化

bug:
後台系統通知功能:新增(目前讀取狀態與通知時間無法正確取值，待解決)

+++ src/main/java/com/cooklab/support_form/model/SFCategory.java
@@ -1,15 +1,15 @@
 package com.cooklab.support_form.model;
 
 public enum SFCategory {
-	RECIPE(1, "食譜問題"),
-	FORUM(2, "討論區問題"),
-    ORDER(3, "訂單問題"),
-	PRODUCT(4, "商品問題");
+	RECIPE((byte) 1, "食譜問題"),
+	FORUM((byte) 2, "討論區問題"),
+    ORDER((byte) 3, "訂單問題"),
+	PRODUCT((byte) 4, "商品問題");
 
-    private final int value;
+    private final Byte value;
     private final String desc;
 
-    SFCategory(int value, String desc) {
+    SFCategory(Byte value, String desc) {
         this.value = value;
         this.desc = desc;
     }
